+++ crates/dioxus-class/src/classes.rs
@@ -1,7 +1,0 @@
-use crate::class::Class;
-
-#[derive(Clone, Debug)]
-pub struct Classes {
-    pub name: String,
-    pub classes: Vec<(String, Class)>,
-}