+++ crates/dioxus-class/src/lib.rs
@@ -3,7 +3,6 @@
 pub use paste;
 
 pub mod class;
-pub mod classes;
 
 pub mod macros;
 
@@ -11,11 +10,6 @@ pub mod build;
 
 pub mod prelude {
     pub use crate::class::Class;
-    pub use crate::classes::Classes;
-
-    pub use crate::class;
-    pub use crate::style_type;
-    pub use crate::style;
 }
 
 pub mod ext {