+++ crates/dioxus-tailwindcss/src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(non_snake_case, non_upper_case_globals)]
 
 pub use dioxus_class;
+pub use dioxus_class_macro;
 
 pub mod macros;
 
@@ -26,6 +27,7 @@ pub mod build;
 
 pub mod prelude {
     pub use dioxus_class::prelude::*;
+    pub use dioxus_class_macro::class;
 
     // macros
     pub use crate::colors;