+++ css.rs
@@ -1,0 +1,112 @@
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(card),
+        String::from(card_compact),
+        String::from(w_64),
+        String::from(h_64),
+        String::from(bg_base_300),
+        String::from(shadow_xl),
+        String::from(text_center),
+        String::from(hover(bg_base_200)),
+        String::from(hover(scale_105))
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![String::from(text_8xl), String::from(py_10)])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(card_body),
+        String::from(text_center),
+        String::from(items_center)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(card_title),
+        String::from(text_sm),
+        String::from(text_base_content)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(flex),
+        String::from(flex_row),
+        String::from(flex_wrap),
+        String::from(justify_center),
+        String::from(gap_4),
+        String::from(mx_2)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(w_80),
+        String::from(h_12),
+        String::from(text_2xl),
+        String::from(m_4),
+        String::from(px_2),
+        String::from(input),
+        String::from(input_primary),
+        String::from(input_bordered)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(bottom_0),
+        String::from(h_full),
+        String::from(mt_20)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![String::from(hidden)])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![String::from(hidden)])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![String::from(w_screen), String::from(h_screen)])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");
+println!(
+    "<div class=\"{}\">",
+    Class::from(vec![
+        String::from(fixed),
+        String::from(top_0),
+        String::from(w_full),
+        String::from(z_10),
+        String::from(bg_base_100),
+        String::from(flex),
+        String::from(justify_center)
+    ])
+);
+println!("{}:{}", "crates/dioxus-css/src/css.rs", 39u32);
+println!("</div>");